move fog extensions from helper to concern
